refactor: refactor responce functions
