+++ async-race/src/modules/features/index.ts
@@ -2,3 +2,5 @@ export { garageMainCreate } from './garageMain';
 export { createCarItem } from './carItem';
 export { winnersMainCreate } from './winnersMain';
 export { createTable } from './winnersTable';
+export * from './responces';
+export * from './buttonsLogic';
